codeDetect/requirements.txt
# Living Documentation Engine - Dependencies
# Install: pip install -r requirements.txt

# Git operations
GitPython>=3.1.40

# Tree-sitter AST parsing (US-6, US-8, US-10)
tree-sitter>=0.21.0
tree-sitter-languages>=1.10.0

# Configuration
PyYAML>=6.0.1

# JSON handling (built-in, listed for documentation)
# json - standard library

# Optional: For enhanced analysis
chardet>=5.2.0  # Encoding detection for binary safety
